requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.10.0
pydantic==2.6.0
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.1.2
sqlalchemy==2.0.25
cryptography==42.0.0
aiosmtplib==3.0.1
email-validator==2.1.0